move filter options to it's own component directory
