+++ src/components/FilterChartOptions/@FilterOptions.scss
@@ -1,0 +1,3 @@
+.filter.well {
+  z-index: 0;
+}
