+++ src/components/FilterChartOptions/FilterBoolean.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react'
 import Select from 'react-select'
-import './@Query.css'
 
 class FilterBoolean extends Component {
   constructor (props) {
